ml/transcriber/requirements.txt
openai-whisper
fastapi
python-multipart
ffmpeg-python
soundfile
numpy
uvicorn
